346: Reorganize the classes in securebanking-openbanking-uk-common
Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/346
