+++ src/test/kotlin/com/forgerock/securebanking/framework/data/DynamicRegistration.kt
@@ -1,6 +1,6 @@
 package com.forgerock.securebanking.framework.data
 
-import com.forgerock.securebanking.openbanking.uk.common.api.meta.OBConstants
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBConstants
 import com.forgerock.uk.openbanking.framework.configuration.OB_SOFTWARE_ID
 import com.forgerock.uk.openbanking.framework.constants.REDIRECT_URI
 import com.forgerock.uk.openbanking.support.discovery.asDiscovery
