+++ src/test/kotlin/com/forgerock/securebanking/framework/signature/SignatureUtils.kt
@@ -3,7 +3,7 @@ package com.forgerock.securebanking.framework.signature
 import com.forgerock.securebanking.framework.configuration.ISS_CLAIM_VALUE
 import com.forgerock.securebanking.framework.data.Tpp
 import com.forgerock.securebanking.framework.utils.GsonUtils
-import com.forgerock.securebanking.openbanking.uk.common.api.meta.OBVersion
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
 import com.forgerock.uk.openbanking.framework.constants.TAN
 import com.forgerock.uk.openbanking.framework.errors.INVALID_DETACHED_JWS_ERROR
 import com.forgerock.uk.openbanking.support.loadRsaPrivateKey
