+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/domestic/payments/consents/GetDomesticPaymentsConsentFundsConfirmationTest.kt
@@ -2,7 +2,7 @@ package com.forgerock.uk.openbanking.tests.functional.payment.domestic.payments.
 
 import assertk.assertThat
 import assertk.assertions.*
-import com.forgerock.securebanking.openbanking.uk.common.api.meta.OBVersion
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
 import com.forgerock.securebanking.framework.conditions.Status
 import com.forgerock.securebanking.framework.configuration.psu
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
